+++ app/Config/Auth.php
@@ -35,7 +35,12 @@ class Auth extends \Myth\Auth\Config\Auth
 	 * @var array
 	 */
 	public $views = [
-		'login'		=> 'App\Views\layouts\login',
+		'login'		   => 'App\Views\layouts\login',
+		'register'		=> 'Myth\Auth\Views\register',
+		'forgot'		  => 'Myth\Auth\Views\forgot',
+		'reset'		   => 'Myth\Auth\Views\reset',
+		'emailForgot'	 => 'Myth\Auth\Views\emails\forgot',
+		'emailActivation' => 'Myth\Auth\Views\emails\activation',
 	];
 
 	/**
@@ -140,10 +145,9 @@ class Auth extends \Myth\Auth\Config\Auth
 	 * When enabled, every registered user will receive an email message
 	 * with an activation link to confirm the account.
 	 *
-	 * @var string Name of the ActivatorInterface class
+	 * @var string|null Name of the ActivatorInterface class
 	 */
-	public $requireActivation = false;
-	//public $requireActivation = 'Myth\Auth\Authentication\Activators\EmailActivator';
+	public $requireActivation = null;
 
 	/**
 	 * --------------------------------------------------------------------
@@ -153,9 +157,9 @@ class Auth extends \Myth\Auth\Config\Auth
 	 * When enabled, users will have the option to reset their password
 	 * via the specified Resetter. Default setting is email.
 	 *
-	 * @var string Name of the ResetterInterface class
+	 * @var string|null Name of the ResetterInterface class
 	 */
-	public $activeResetter = 'Myth\Auth\Authentication\Resetters\EmailResetter';
+	public $activeResetter = null;
 
 	/**
 	 * --------------------------------------------------------------------